+++ Tests/SwiftLintFramework/RulesTests.swift
@@ -233,6 +233,10 @@ class RulesTests: XCTestCase {
         verifyRule(SwitchCaseOnNewlineRule.description)
     }
 
+    func testSyntacticSugar() {
+        verifyRule(SyntacticSugarRule.description)
+    }
+
     func testTodo() {
         verifyRule(TodoRule.description, commentDoesntViolate: false)
     }